use a map instead of list for notifications, inspiration from docker-compose and github actions
